requirements.txt
flask==1.0.2
pandas
numpy
requests
scipy
mysql-connector-python